Major bug fix
Fixed a bug which would sometimes set the quality factor for the resonance compensation to 1. Fix was implemented by hardcoding the value 50 for the quality factor. In tests, we have measured it to be higher, but 50 is the max setting for the KSC-2, which makes sense considering the fact that our resonator is unnaturally good at resonating.